Combat works!
combat works.  If you're ranged and farther than 1 tile away you won't get counter attacked but will if you're right next to the guy you're attacking.  It now also checks if someone is dead after the initial attack so that if you die on the first hit you don't get to counter attack.  And units die when they are killed