Added full cursor functionality
I changed the OnMouseDown and OnMouseEnter functions to new functions.  In the update function I made it so if the cursor object is on top of an allied unit and you press space bar it will call the selected function.  Or if you press C it'll call the function that brings up your stats.  I also made it so you can select the tile you want to go to with the cursor as well.  I changed the OnMouseDown function in the tile class to a function that checks if you pressed the space bar while the cursor object is on top of a tile that is within range.  I also went ahead and deleted all the OnMouseDown functions